Rebased to use parseAdditional code
Ethan and Nik had worked on extending Query and other classes to use the OPTRR work Andrew and I had worked on
we needed to merge our code together and this was best accomplished using a rebase.

The pom.xml was also updated to use a new version of the maven compile plugin as there was a bug stopping me from
being able to build the project.